Add TagCommand feature to filter staff contacts based on tag
This feature allows for users to search for staff contacts based on the tag associated with the specified contacts. Tag command is case insensitive.

Implementing this will make it easier for users to find contacts under a specific tag. A case insensitive tag feature is more user-friendly because users cannot be
expected to remember the exact case of the keywords.

Let's,
* update the TagCommand algorithm to use case-insensitive matching
* parse TagCommand